+++ api/app.go
@@ -1,4 +1,4 @@
-package vonji
+package api
 
 import (
 	"github.com/gorilla/mux"
