+++ api/context.go
@@ -1,4 +1,4 @@
-package vonji
+package api
 
 import (
 	"github.com/jinzhu/gorm"
